Create zip_file_cracker.py
The program takes an input of a word list to guess the passwords with and a zip file to target. The program finishes by printing out the answer if it turns to be the zip file's password. Else, it prints that the word list didn't work. Try it and see if there are any problems with the code above.